flask_app/requirements.txt
leancloud-sdk
Flask>=0.10.1
ndg-httpsclient
pyasn1
flask-cors